requirements.txt
Flask==3.0.0
openai==1.51.0
azure-storage-blob==12.19.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
openpyxl==3.1.2
xlsxwriter==3.1.9
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.11.0
